+++ tests/test_model_hofstadter.py
@@ -1,10 +1,10 @@
 # Copyright 2018 TeNPy Developers
 from tenpy.models.hofstadter import HofstadterBosons, HofstadterFermions
 from test_model import check_general_model
-from nose.plugins.attrib import attr
+import pytest
 
 
-@attr('slow')
+@pytest.mark.slow
 def test_HofstadterBosons():
     model_pars = {
         'Lx': 3,
@@ -26,7 +26,7 @@ def test_HofstadterBosons():
     check_general_model(HofstadterBosons, model_pars, {'bc_MPS': ['finite', 'infinite']})
 
 
-@attr('slow')
+@pytest.mark.slow
 def test_HofstadterFermions():
     model_pars = {'Lx': 3, 'Ly': 3, 'phi': (1, 3), 'conserve': 'N', 'v': 0.456, 'mu': 0.123}
     check_general_model(HofstadterFermions, model_pars, {